pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vezgo"
version = "1.0.0"
description = "Official Vezgo Python SDK for Cryptocurrency Data API"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Vezgo", email = "hello@vezgo.com"}
]
maintainers = [
    {name = "Vezgo", email = "hello@vezgo.com"}
]
keywords = [
    "vezgo",
    "cryptocurrency",
    "crypto",
    "api",
    "sdk",
    "bitcoin",
    "ethereum",
    "coinbase",
    "binance",
    "blockchain",
    "wallet",
    "exchange",
    "portfolio",
    "fintech"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Financial",
]
requires-python = ">=3.8"
dependencies = [
    "httpx>=0.24.0",
    "pyjwt>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://vezgo.com"
Documentation = "https://vezgo.com/docs"
Repository = "https://github.com/wealthica/vezgo-sdk-python"
Changelog = "https://github.com/wealthica/vezgo-sdk-python/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

